# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Campaign Master API
  version: "1.0"
  description: |
    REST API for Campaign Master covering authentication, wallets, marketplace,
    ad serving, tracking, and reports.
servers:
  - url: http://localhost:5001
paths:
...
  /api/adserve:
    get:
      summary: Serve ad
      parameters:
        - name: slot_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: HTML snippet with ad creative
          content:
            text/html:
              schema:
                type: string
        '204':
          description: No eligible ad
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/track/impression:
    get:
      summary: Track impression
      parameters:
        - name: request_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Tracked or already recorded
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Reservation expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/track/click:
    get:
      summary: Track click
      parameters:
        - name: request_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirect to landing URL
          headers:
            Location:
              schema:
                type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Reservation expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
  responses:
    BadRequest:
      description: Invalid request body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...